Fix URLClassLoader bug
In Java 9+ ClassLoader is not an instance of URLClassloader.

I'm not sure why the classloader was used to get the classpath here -
I'm not 100% sure that the code isn't loading jars dynamically but I
don't understand why it would.

That being said, this code seems to be unused in practice other than
tests and so once I get the tests passing I'll try to remove it
entirely.