ModuleTesterTest.testPythonModuleExample passes
so does testPythonModuleError as a bonus

Steps taken
* Commented out the dockerfile editing as it was trying to access files
in /kb/dev_container which doesn't exist in sdkbase2
* Found that nose failed repeatedly with the error
```
  File "/miniconda/lib/python3.6/site-packages/nose/case.py", line 346,
in __init__
    self.inst = self.cls()
TypeError: __init__() missing 1 required positional argument: 'config'
```
* While trying various things, found that reverting the python templates
to an earlier version of the example code made the tests pass. Commit of
reversion is 0391723. Once tests are passing we can look into restoring
the example code.
  * Updated the old examples to use the new location for the workspace
client.