Remove `compile` `--javalib` and `--javabuildxml` args
Unused in kbase and makes switching to gradle much simpler since we
don't need to worry about how the hard coded jars in the
`JavaTypeGenerator` interact with the build system. Now we just have to
worry about 1 test file which should be a lot simpler and I think will
Just Work (TM).