Add and use a TestLocalManager class
Ensures test_local is set up correctly when init, test, or validate is
run, and that an appropriate error message is provided for test if test
credentials are missing.

Fixes a regression that caused the test_local directory to not be
regenerated if it's missing on kb-sdk test; would return an unhelpful
message that the directory wasn't set up correctly.