Update the ModuleRunner to use the CallbackServerManager.
There's no tests for the MR so this was tested manually.

The mount option was removed from the run command for two reasons:

* The Python callback server doesn't support user specified mounts
* User mounts doesn't necessarily work. If a SDK module tries to
hardlink an input file from a user mount to the scratch directory, it
will fail as that appears to be a inter-device mount to the docker
image. The only way to ensure that hardlinks work is by putting the
input files in the scratch directory, which makes them visible to all
other containers as well.

The drawback is that we can no longer safely clean up the workdir
directory as it may contain input files from the user.