Fix a bug that prevents java compilation
... unless a client is requested.

This bug was introduced when removing code that forced client
compilation when the server was requested. It apparently was not
detected because the tests run in docker containers where the old kb-sdk
recompiles the spec. It was discovered when replacing the base image and
using the new graalvm compiled binary.